src/services/csv_handler.py
import pandas as pd

def read_csv(file_path):
    """Reads a CSV file and returns a DataFrame."""
    try:
        data = pd.read_csv(file_path)
        return data
    except Exception as e:
        print(f"Error reading CSV file: {e}")
        return None

def process_csv_data(data):
    """Processes the DataFrame and returns cleaned data."""
    if data is not None:
        # Example processing: drop rows with any missing values
        cleaned_data = data.dropna()
        return cleaned_data
    return None